Improve efficiency of status
Instead of hitting the 'extended' status api for every combination of file and locale, use the file status endpoint for each file.  The payload from this endpoint contains information for each locale.  This reduces the number of requests from: number of locales * number of file to simply the number of files.